Added a couple safety checks and improved output formatting

+++ smol_build.sh
@@ -2,12 +2,31 @@
 set -euo pipefail
 IFS=$'\n\t'
 
-echo "smol build is smol building..."
-echo
+# Check that two arguments were given
+if [ $# -ne 2 ]; then
+    echo "Error: Two arguments required!" >&2
+    echo "Usage: $0 <source_dir> <build_dir>" >&2
+    echo "Warning: The build directory will be emptied when the script is run." >&2
+    exit 1
+fi
 
-# TODO - make directory usage robust
 source_dir=$1;
 build_dir=$2;
+
+# Make sure the build directory isn't the current directory
+if [ "$build_dir" = "." ]; then
+    echo "Error: Build directory cannot be current directory (.)" >&2
+    echo "Warning: The build directory will be emptied when the script is run." >&2
+    exit 1
+fi
+
+if [ ! -d "$source_dir" ] ; then
+    echo "Error: Source input must be a directory" >&2
+    exit 1
+fi
+
+echo "smol build is building..."
+echo
 
 # Create (if needed) and then clear out the build directory
 mkdir -p $2 && find $2 -mindepth 1 -depth -exec rm -rf {} +
@@ -16,7 +35,7 @@ descend() { # param1: recursive directory | param2: template_string_so_far
 
 	local template=$2
 	if [[ -f "$1/smol_template.html" ]]; then
-		echo " [tmplt] $1/smol_template.html"
+		echo " [templating] $1/smol_template.html"
 		if [[ -z "$template" ]]; then
 		    template=$(cat $1/smol_template.html)
 	    else
@@ -31,14 +50,14 @@ descend() { # param1: recursive directory | param2: template_string_so_far
     	elif [ -f "$item" ] && [[ "$item" =~ \.html$ ]]; then
     		if [[ ! "$item" =~ ^.*smol_template\.html$ ]]; then
 			    local output_path="$build_dir/${item#*/}"
-			    echo " [build] $item -> $output_path"
+			    echo " [building]   $item -> $output_path"
 		    	output_data="${template//<smol_content\/>/$(cat $item)}"
 			    mkdir -p $(dirname "$output_path")
 		    	printf '%s' "$output_data" > "$output_path"
 	    	fi
 		else # All other static files
 			local output_path="$build_dir/${item#*/}"
-			echo " [copy]  $item -> $output_path"
+			echo " [copying]    $item -> $output_path"
 			mkdir -p $(dirname "$output_path")
 			cp $item $output_path
         fi
@@ -46,7 +65,5 @@ descend() { # param1: recursive directory | param2: template_string_so_far
 }
 
 descend $source_dir ""
-
-# TODO - quite mode
-# TODO - Handle CSS is special case
-# TODO - Handle static files
+echo
+echo "smol build is done!"
